chore: open local folder when fail to init with remote server data
